# 04-Source-code/docs/api/v1/openapi.yaml
...
openapi: 3.0.3
info:
  title: Alarmsystem-Backend G2 — API v1 (Vereisungserkennung ANR)
  version: 1.0.0
  description: |
    REST-Schnittstelle des G2-Backends (Erfassung & Bewertung von
    Vereisungsbedingungen am Regionalflughafen ANR). **G2 ist hier Server**,
    **G3 (Frontend) ist Client** (GET / SSE-Abo — G3 hostet nichts).

    **Sicherheits-Randbedingung RB-01:** Das System ist reine
    *Entscheidungsunterstuetzung*. Es gibt **keinen** Endpoint, der die Startbahn
    freigibt/sperrt/steuert. Quittieren eines Alarms ist eine reine UI-/Audit-
    Aktion, kein Steuerbefehl.

    **Fail-safe NF-01:** Bei veralteten (stale) oder fehlenden Daten liefert die
    aktuelle Bewertung **nie GRUEN**, sondern `risk_level=unknown` +
    `is_stale=true` mit HTTP `200` — das ist **kein** Fehlerfall. `503` nur, wenn
    G2 selbst nicht lieferfaehig ist (z. B. noch keine Daten / interner Ausfall).

    Quelle: `docs/API_FROZEN_v1.md`. Konsumierter G1-Vertrag: siehe
    `g1-consumed.openapi.yaml`.
servers:
  - url: http://localhost:8000
    description: Lokale Entwicklung (Prototyp)
...
paths:
...
  /v1/thresholds:
    get:
      tags: [Thresholds]
      operationId: getThresholds
      summary: Aktuelle Schwellenwerte lesen (DTB-62)
      description: |
        Liefert die aktuell konfigurierten Vereisungs-Schwellenwerte (aus
        `config/thresholds.json` via Loader DTB-15) fuer das G3-Schwellenwert-Menue.
        Werte sind parametrierbar (NF-05); aktuell DUMMY (G1-Finalwerte folgen).

        **RB-01-neutral:** rein lesend, kein Aktor. **Aendern** der Schwellen erfolgt
        spaeter ueber einen separaten, Auth-geschuetzten Endpoint (NF-07).

        > Nach dem v1.0-Freeze ergaenzt (DTB-62) — Naht-Review durch Architekt.
      responses:
        "200":
          description: Aktuelle Schwellenwerte.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thresholds"
              example:
                vereisung:
                  t_s_gefrierpunkt_c: 0.0
                  t_s_gelb_auffang_c: 1.0
                  delta_t_kondensation_k: 0.0
                  delta_t_feucht_k: 1.0
                prognose:
                  t_s_grenz_c: 0.0
                datenqualitaet:
                  stale_timeout_s: 120
                  max_temp_jump_c_per_min: 5.0
                  flatline_timeout_min: 15.0
                  flatline_epsilon_c: 0.15
                  max_clock_skew_s: 5.0
                  min_plausible_dew_point_c: -50.0
                plausibilitaet:
                  min_temp_c: -50.0
                  max_temp_c: 50.0
                  min_humidity_pct: 0.0
                  max_humidity_pct: 100.0
                  min_pressure_hpa: 800.0
                  max_pressure_hpa: 1100.0
        "503":
          description: >-
            G2 (noch) nicht lieferfaehig — Runtime nicht bereit (z. B. Schwellen-Config
            beim Start nicht ladbar). Generische Meldung, keine internen Details (Contract D).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: SERVICE_UNAVAILABLE
                message: "G2 momentan nicht lieferfaehig."
  /v1/assessment/current:
    get:
      tags: [Assessment]
      operationId: getCurrentAssessment
      summary: Aktuelle Vereisungsbewertung
      description: |
        Liefert die zuletzt berechnete Bewertung als **Ampel + Roh-Messwerte**
        in einem Response (flach, kein Envelope — E-36).

        **Fail-safe (NF-01):** Sind die zugrunde liegenden Daten veraltet
        (stale) oder fehlt G1, ist dies **kein** Fehler: Response `200` mit
        `is_stale=true` und `risk_level=unknown` (nie GRUEN). `503` nur, wenn G2
        selbst (noch) nicht lieferfaehig ist.

        **RB-01:** rein beratend — keine Freigabe-/Steueraktion.
      responses:
        "200":
          description: Aktuelle Bewertung (ggf. Fail-safe `unknown` + `is_stale=true`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentCurrent"
              examples:
                normal:
                  summary: Normalfall (Risiko gelb)
                  value:
                    risk_level: yellow
                    driving_factor: dew_point
                    explanation: "ΔT ≤ 1,0 °C an der Oberfläche → Feuchte"
                    surface_temp_c: -0.4
                    dew_point_c: -1.1
                    delta_t: 0.7
                    humidity_pct: 96
                    measured_at: "2026-06-22T14:03:05Z"
                    assessed_at: "2026-06-22T14:03:30Z"
                    is_stale: false
                    sensor_status: ok
                failsafe_stale:
                  summary: Fail-safe — veraltete Daten (NF-01)
                  value:
                    risk_level: unknown
                    driving_factor: null
                    explanation: "Letzter Messwert älter als der Stale-Timeout — keine gültige Bewertung."
                    surface_temp_c: null
                    dew_point_c: null
                    delta_t: null
                    humidity_pct: null
                    measured_at: "2026-06-22T13:55:00Z"
                    assessed_at: "2026-06-22T14:03:30Z"
                    is_stale: true
                    sensor_status: ok
                failsafe_fault:
                  summary: Fail-safe — Sensor defekt trotz aktueller Daten (NF-01, nie GRUEN)
                  value:
                    risk_level: unknown
                    driving_factor: null
                    explanation: "Sensor meldet 'fault' — keine vertrauenswürdige Bewertung möglich."
                    surface_temp_c: null
                    dew_point_c: null
                    delta_t: null
                    humidity_pct: null
                    measured_at: "2026-06-22T14:03:05Z"
                    assessed_at: "2026-06-22T14:03:30Z"
                    is_stale: false
                    sensor_status: fault
        "503":
          description: G2 ist nicht lieferfaehig (z. B. noch keine Daten / interner Ausfall). NICHT fuer Stale.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: SERVICE_UNAVAILABLE
                message: "Noch keine Bewertung verfügbar."
...
  /v1/alarms/{id}/ack:
    post:
      tags: [Alarms]
      operationId: acknowledgeAlarm
      summary: Alarm quittieren (UI-/Audit-Aktion, kein Aktor)
      description: |
        Quittiert einen Alarm. **RB-01:** rein dokumentierend (Audit) — die UI
        kann den Alarm damit beenden, aber dies ist **kein** Bahn-Freigabe-/
        Sperr-Befehl. Setzt den Alarm-Zustand auf `acknowledged`.

        **Double-Ack (NF-09, Audit-Integritaet):** Ist der Alarm bereits
        `acknowledged` oder `cleared`, wird eine erneute Quittierung mit `409`
        abgelehnt (nicht idempotent, kein lautloses Schlucken) — so bleibt jede
        Quittierung ein eindeutiger, nachvollziehbarer Audit-Eintrag.

        **Auth (DTB-63, NF-07):** Erfordert einen gueltigen `X-API-Key`-Header
        (statischer API-Key, serverseitig aus `G2_API_KEY`). Fehlender/falscher
        Schluessel -> `401`. Reiner Zugangsschutz — RB-01 unberuehrt (kein Aktor).
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID des zu quittierenden Alarms.
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AckRequest"
            example:
              operator: "tower-ops-01"
              note: "Gesehen, manuelle Sichtkontrolle eingeleitet."
      responses:
        "200":
          description: Quittierung gespeichert (append-only Audit-Eintrag).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Acknowledgement"
        "401":
          description: Fehlender oder ungueltiger API-Key (DTB-63, NF-07).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: UNAUTHORIZED
                message: "Ungueltiger oder fehlender API-Key."
        "400":
          description: "Ungueltiger Pfad-Parameter `id` (z. B. `id=0`, verletzt `minimum: 1`)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: BAD_REQUEST
                message: "Ungueltige Alarm-ID."
        "404":
          description: Alarm mit dieser ID existiert nicht.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: NOT_FOUND
                message: "Alarm 42 nicht gefunden."
        "409":
          description: Alarm ist bereits quittiert/geschlossen — erneute Quittierung abgelehnt (NF-09).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: ALARM_ALREADY_ACKNOWLEDGED
                message: "Alarm 42 ist bereits im Zustand 'acknowledged'."
        "422":
          description: |
            Ungueltiger Request-Body (Body-Schema-Validierung, z. B. `operator`
            fehlt). Konvention: `400` = Request-/Pfad-/Geschaeftsregel-Fehler (wie
            bei `/v1/readings` und `/v1/alarms`); `422` = Body-Schema-Validierung.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "503":
          description: Schreibzugriff nicht konfiguriert (G2_API_KEY fehlt) oder G2 nicht lieferfaehig.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: SERVICE_UNAVAILABLE
                message: "Schreibzugriff nicht konfiguriert."
  /v1/config:
    post:
      tags: [Thresholds]
      operationId: updateThresholds
      summary: Schwellenwerte aktualisieren (DTB-63)
      description: |
        Schreibt einen neuen Satz parametrierbarer Schwellenwerte. Erfordert einen
        gueltigen `X-API-Key`-Header (DTB-63, NF-07). Die neue Config wird in
        `config/thresholds.json` gespeichert und die Runtime neu geladen, sodass
        `GET /v1/thresholds` und die Bewertungslogik sofort die neuen Werte verwenden.

        RB-01: reine Konfiguration, keine Steuer-Aktion.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Thresholds"
      responses:
        "200":
          description: Gespeicherte Schwellenwerte.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thresholds"
        "401":
          description: Fehlender oder ungueltiger API-Key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: UNAUTHORIZED
                message: "Ungueltiger oder fehlender API-Key."
        "422":
          description: Ungueltige Schwellenwerte (Body-Validierung).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: UNPROCESSABLE_ENTITY
                message: "Config-Inkonsistenz: ..."
        "503":
          description: G2 nicht lieferfaehig (Config nicht schreibbar / Runtime-Reload fehlgeschlagen).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: SERVICE_UNAVAILABLE
                message: "G2 momentan nicht lieferfaehig."
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Statischer API-Key fuer **schreibende** Endpoints (DTB-63, NF-07). Soll-Wert
        serverseitig aus Umgebungsvariable `G2_API_KEY` (kein Secret im Code). Lesende
        Endpoints sind offen. Methode bewusst wechselbar (Basic/JWT spaeter) —
        Naht-Entscheidung beim Architekt.
  schemas:
...
    RiskLevel:
      type: string
      description: |
        Vereisungs-Risikostufe. `unknown` ist der Fail-safe-Zustand (NF-01):
        bei veralteten/fehlenden Daten nie GRUEN.
      enum: [green, yellow, orange, red, unknown]
    SensorStatus:
      type: string
      description: Sensor-/Lieferstatus (aus G1s `GET /current` `status`).
      enum: [ok, fault]
...
    Thresholds:
      type: object
      description: |
        Aktuell konfigurierte Vereisungs-Schwellenwerte (Spiegel von
        config/thresholds.json via Loader DTB-15). Werte sind parametrierbar
        (NF-05); aktuell DUMMY (G1-Finalwerte folgen).
      required: [vereisung, prognose, datenqualitaet, plausibilitaet]
      properties:
        vereisung:
          type: object
          required:
            [t_s_gefrierpunkt_c, t_s_gelb_auffang_c, delta_t_kondensation_k, delta_t_feucht_k]
          properties:
            t_s_gefrierpunkt_c:
              type: number
              description: Oberflaechentemp-Gefrierpunkt (°C).
            t_s_gelb_auffang_c:
              type: number
              description: Obere GELB-Auffang-Grenze fuer T_s (°C).
            delta_t_kondensation_k:
              type: number
              description: ΔT-Schwelle Kondensation (K).
            delta_t_feucht_k:
              type: number
              description: ΔT-Schwelle "Feuchte vorhanden" (K).
        prognose:
          type: object
          required: [t_s_grenz_c]
          properties:
            t_s_grenz_c:
              type: number
              description: T_s-Grenze fuer die GELB-Prognoseschwelle (°C).
        datenqualitaet:
          type: object
          required:
            [stale_timeout_s, max_temp_jump_c_per_min, flatline_timeout_min, flatline_epsilon_c, max_clock_skew_s, min_plausible_dew_point_c]
          properties:
            stale_timeout_s:
              type: number
              description: Timeout nach dem ein Messwert als veraltet gilt (s).
            max_temp_jump_c_per_min:
              type: number
              description: Maximal zulaessiger Temperatursprung pro Minute (°C/min).
            flatline_timeout_min:
              type: number
              description: Timeout fuer erkannte Flatline (min).
            flatline_epsilon_c:
              type: number
              description: Toleranz fuer Flatline-Erkennung (°C).
            max_clock_skew_s:
              type: number
              description: Maximal zulaessige Uhrzeitabweichung (s).
            min_plausible_dew_point_c:
              type: number
              description: Minimal plausibler Taupunkt (°C).
        plausibilitaet:
          type: object
          required:
            [min_temp_c, max_temp_c, min_humidity_pct, max_humidity_pct, min_pressure_hpa, max_pressure_hpa]
          properties:
            min_temp_c:
              type: number
              description: Minimale plausible Temperatur (°C).
            max_temp_c:
              type: number
              description: Maximale plausible Temperatur (°C).
            min_humidity_pct:
              type: number
              description: Minimale plausible Luftfeuchte (%).
            max_humidity_pct:
              type: number
              description: Maximale plausible Luftfeuchte (%).
            min_pressure_hpa:
              type: number
              description: Minimaler plausibler Luftdruck (hPa).
            max_pressure_hpa:
              type: number
              description: Maximaler plausibler Luftdruck (hPa).

    # --- Aktuelle Bewertung (G2 -> G3), flach (E-36) ---
    AssessmentCurrent:
      type: object
      description: |
        Aktuelle Vereisungsbewertung. Felder gemaess `Assessment`-Schema
        (src/model/schemas.py): `assessed_at` entspricht dort dem Feld `ts`;
        `measured_at` stammt aus dem verknuepften `Reading` (G1-Snapshot, nicht
        aus `Assessment`). Naht-Felder `is_stale`/`sensor_status` ergaenzen die
        Naht-Sicht. Alle Zeitstempel in UTC (ISO-8601).

        **Fail-safe-Invarianten (NF-01) — verbindlich, von der `assessment/`-Logik
        garantiert (OpenAPI 3.0.3 kann felduebergreifende Regeln nicht hart
        erzwingen; Durchsetzung + Tests liegen im kritischen Pfad):**
        - `risk_level=green` ist NUR zulaessig, wenn `is_stale=false` UND
          `sensor_status=ok`. Bei `is_stale=true` oder `sensor_status=fault`
          gilt `risk_level=unknown` (nie GRUEN).
        - `measured_at` ist auf einer 200-Antwort **immer** gesetzt (auch bei
          `risk_level=unknown`); der Fall „gar keine Daten" wird mit **HTTP 503**
          signalisiert, nicht mit `null`.
        - Bei `risk_level != unknown` sind `surface_temp_c`, `dew_point_c`,
          `delta_t` und `humidity_pct` nicht-null (eine echte
          Bewertung hat immer Messwerte). Genullte Messwerte treten **nur** bei
          `risk_level=unknown` auf — ausgeloest durch `is_stale=true` (veraltet)
          **oder** `sensor_status=fault` (Sensor defekt). `measured_at` kann dabei
          im Fault-Fall mit frischen Daten gesetzt sein (Zeitstempel bekannt, Werte
          aber nicht vertrauenswuerdig); bei reinem Stale traegt es den letzten,
          alten Zeitstempel.
      required: [risk_level, assessed_at, is_stale, sensor_status, measured_at]
      properties:
        risk_level:
          $ref: "#/components/schemas/RiskLevel"
        driving_factor:
          type: string
          nullable: true
          maxLength: 64  # = VARCHAR(64) in migrations/schema.sql (assessment.driving_factor)
          description: Hauptfaktor der Bewertung (z. B. `dew_point`, `surface_temp`).
        explanation:
          type: string
          nullable: true
          maxLength: 512  # = VARCHAR(512) in migrations/schema.sql (assessment.explanation)
          description: Menschenlesbare Begruendung der Stufe.
        surface_temp_c:
          type: number
          nullable: true
          description: Oberflaechentemperatur T_s (°C).
        dew_point_c:
          type: number
          nullable: true
          description: Taupunkt (°C), von G2 berechnet (Magnus).
        delta_t:
          type: number
          nullable: true
          description: ΔT = surface_temp_c − dew_point_c (°C).
        humidity_pct:
          type: number
          nullable: true
          description: Luftfeuchte (%).
        measured_at:
          type: string
          format: date-time
          description: |
            G1-Messzeitpunkt des zugrundeliegenden Readings (UTC). Auf einer 200-Antwort
            immer gesetzt (bei `is_stale=true` = letzter bekannter Zeitstempel). Der Fall
            „gar keine Daten vorhanden" wird mit HTTP 503 signalisiert, nicht mit `null`.
        assessed_at:
          type: string
          format: date-time
          description: G2-Bewertungszeitpunkt (UTC).
        is_stale:
          type: boolean
          description: "`true` = Daten ueberaltert; zusammen mit `risk_level=unknown` = Fail-safe griff (NF-01)."
        sensor_status:
          $ref: "#/components/schemas/SensorStatus"

    # --- Messwert (Historie) ---
...
    AckRequest:
      type: object
      description: Request-Body fuer die Quittierung eines Alarms.
      required: [operator]
      properties:
        operator:
          type: string
          minLength: 1  # leer waere kein gueltiger Audit-Eintrag (NF-09)
          maxLength: 128  # = VARCHAR(128) in migrations/schema.sql (acknowledgement.operator)
          description: Kennung der quittierenden Person (Audit).
        note:
          type: string
          nullable: true
          maxLength: 512  # = VARCHAR(512) in migrations/schema.sql (acknowledgement.note)
          description: Optionale Notiz.
    Acknowledgement:
      type: object
      description: Gespeicherte Quittierung (append-only, NF-09).
      required: [id, alarm_id, operator, ts]
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        alarm_id:
          type: integer
          format: int64
        operator:
          type: string
        note:
          type: string
          nullable: true
        ts:
          type: string
          format: date-time
          description: Zeitpunkt der Quittierung (UTC).

    # --- Fehlerformat ---
    Error:
      type: object
      description: Maschinenlesbarer Fehler (Code + Meldung). Keine internen Details/Secrets.
      required: [code, message]
      properties:
        code:
          type: string
          description: Maschinenlesbarer Fehlercode (z. B. BAD_REQUEST, NOT_FOUND, SERVICE_UNAVAILABLE).
        message:
          type: string
          description: Menschenlesbare Fehlermeldung.